+++ index.html
@@ -235,7 +235,7 @@
     </section>
   </template>
 
-  <script src="js/main.js" defer></script>
+  <script src="js/main.js" type="module"></script>
 
 </body>
